+++ crates/aranya-daemon/src/sync/mod.rs
@@ -1,4 +1,5 @@
 //! Aranya syncer for syncing Aranya graph commands.
 
+pub mod quic;
 pub mod task;
-pub mod tcp;
+pub mod prot;
