+++ crates/aranya-daemon/src/sync/prot.rs
@@ -1,0 +1,8 @@
+//! Aranya syncer supported protocols.
+//! New protocols must be added to the end of this list since protocol types can never change.
+
+/// Protocols supported by the syncer.
+pub enum SyncProtocols {
+    /// QUIC syncer protocol.
+    QUIC,
+}
